requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.53.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
html2text==2024.2.26
httplib2==0.22.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
isort==5.13.2
jedi==0.19.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.2
jupyter-core==5.7.2
kiwisolver==1.4.5
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-experimental==0.0.62
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2
langsmith==0.1.82
markdown==3.6
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
pyzmq==26.0.3
requests==2.32.3
rsa==4.9
six==1.16.0
sqlalchemy==2.0.31
stack-data==0.6.3
tenacity==8.4.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4
